Move functionality from 'global' module to 'core'
