+++ modules/core/server-scala/src/main/scala/common/graphql/schema/GraphQLSchema.scala
@@ -1,0 +1,7 @@
+package common.graphql.schema
+
+import common.shapes.ServerModule
+
+class GraphQLSchema(module: ServerModule) extends GraphQL {
+  override val serverModule: ServerModule = module
+}
