--- packages/server-scala/src/main/scala/app/GlobalModule.scala
@@ -1,5 +1,0 @@
-package app
-
-import common.shapes.ServerModule
-
-class GlobalModule(modules: Seq[ServerModule]) extends ServerModule(modules)
