+++ packages/server-scala/src/main/scala/guice/ServerModuleBinding.scala
@@ -2,14 +2,22 @@ package guice
 
 import app._
 import com.google.inject.{Provides, Singleton}
-import common.graphql.schema.GraphQL
-import graphql.GraphQLSchema
+import common.shapes.ServerModule
+import core.guice.bindings.{CoreBinding, SangriaBinding}
 import net.codingwell.scalaguice.ScalaModule
 
-class ServerModulesBinding extends ScalaModule {
+class ServerModuleBinding extends ScalaModule {
 
   override def configure(): Unit = {
-    bind(classOf[GraphQL]).to(classOf[GraphQLSchema])
+    install(new ContactBinding)
+    install(new CoreBinding)
+    install(new CounterBinding)
+    install(new MailBinding)
+    install(new ItemBinding)
+    install(new UserBinding)
+    install(new FileBinding)
+
+    install(new SangriaBinding)
   }
 
   @Provides
@@ -18,8 +26,8 @@ class ServerModulesBinding extends ScalaModule {
                     userModule: UserModule,
                     uploadModule: UploadModule,
                     paginationModule: PaginationModule,
-                    counterModule: CounterModule): GlobalModule = {
-    new GlobalModule(
+                    counterModule: CounterModule): ServerModule = {
+    new ServerModule(
       Seq(
         contactModule,
         userModule,
